+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Disable last-applied-configuration annotation for replaced CRDs (https://github.com/pulumi/pulumi-kubernetes/pull/1868)
 - Fix Provider config diffs (https://github.com/pulumi/pulumi-kubernetes/pull/1869)
 - Fix replace for named resource using server-side diff (https://github.com/pulumi/pulumi-kubernetes/pull/1870)
+- Fix import for Provider using server-side diff (https://github.com/pulumi/pulumi-kubernetes/pull/1872)
 
 ## 3.14.0 (January 12, 2022)