+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD (Unreleased)
 
+## 3.14.1 (January 18, 2022)
+
 - Disable last-applied-configuration annotation for replaced CRDs (https://github.com/pulumi/pulumi-kubernetes/pull/1868)
 - Fix Provider config diffs (https://github.com/pulumi/pulumi-kubernetes/pull/1869)
 - Fix replace for named resource using server-side diff (https://github.com/pulumi/pulumi-kubernetes/pull/1870)